requirements.txt
azure-core==1.32.0
azure-storage-blob==12.24.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
pycparser==2.22
python-dotenv==1.0.1
pydantic==2.10.6
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
